requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
email-validator
pydantic-settings
python-dotenv
openai>=1.0
anthropic
google-generativeai
gunicorn>=20.1.0
streamlit
requests
loguru
sqlalchemy
alembic
psycopg2-binary

# --- 认证依赖 ---
# Pin versions to avoid bcrypt 4.x incompatibility with passlib <1.7.5
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-jose[cryptography]>=3.3.0
python-multipart
psycopg2-binary
rich
volcengine-python-sdk[ark]